refactor(ChatRoom): actualizar la importación del componente MessageBubble a su nueva ubicación en el directorio de mensajes
