+++ hooks/db/useChatsDb.ts
@@ -2,7 +2,7 @@ import { useState, useEffect, useCallback } from 'react';
 import { db } from '../../database/db';
 import { chats, chatParticipants, messages } from '../../database/schema';
 import { eq } from 'drizzle-orm';
-import { MessageStatus } from '@/components/MessageStatus'
+import { MessageStatus } from '@/components/messages/MessageStatus'
 
 export interface Message {
   id: string;
